requirements.txt
asgiref==3.8.1
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
Flask==2.3.3
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
mysqlclient==2.2.7
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipenv==2025.0.2
platformdirs==4.3.8
protobuf==4.21.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.3
setuptools==80.4.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
